+++ bin/demo.mjs
@@ -1,0 +1,4 @@
+import shell from 'shelljs';
+import fs from 'node:fs/promises';
+
+async function main() {}