+++ packages/toast/src/components/base-toast/component.tsx
@@ -12,13 +12,14 @@ import cn from 'classnames';
 
 import { Popover, PopoverProps } from '@alfalab/core-components-popover';
 import { Portal } from '@alfalab/core-components-portal';
+import { noop } from '@alfalab/core-components-shared';
 import { Stack } from '@alfalab/core-components-stack';
+import { stackingOrder } from '@alfalab/core-components-stack-context';
 import {
     ToastPlate as ToastPlateComponent,
     ToastPlateProps,
 } from '@alfalab/core-components-toast-plate';
 import { useClickOutside, usePrevious } from '@alfalab/hooks';
-import { stackingOrder } from '@alfalab/stack-context';
 
 import styles from './index.module.css';
 
@@ -87,8 +88,6 @@ export type BaseToastProps = ToastPlateProps &
         offset?: [number, number];
     };
 
-const noop = () => {};
-
 export const BaseToast = forwardRef<HTMLDivElement, BaseToastProps>(
     (
         {
@@ -123,7 +122,9 @@ export const BaseToast = forwardRef<HTMLDivElement, BaseToastProps>(
         const startTimer = useCallback(() => {
             clearTimeout(timerId.current);
 
-            timerId.current = window.setTimeout(onClose, autoCloseDelay);
+            timerId.current = window.setTimeout(() => {
+                onClose?.();
+            }, autoCloseDelay);
         }, [autoCloseDelay, onClose]);
 
         const stopTimer = useCallback(() => {