+++ packages/side-panel/CHANGELOG.md
@@ -1,4 +1,4 @@
-# Change Log
+# @alfalab/core-components-side-panel
 
 ## 5.9.11